project: reorganize code
This commit renames the core package to 'patchtools' and lower cases the
individual module names as is more typical in python.

We also move the documentation to doc/ and the scripts to scripts/, which
simplifies setup.py.  patch.cfg is also not installed by default and is only
provided as an example since it is pretty useless otherwise.

In addition, exportpatch and fixpatch will now use sane defaults in the
absence of a config file.