patchtools: use subprocess.run instead of subprocess.Popen
Python 3.5 introduced subprocess.run as an easier way to handle
subprocess communication.  Popen() stopped accepting the 'text'
argument so it's a good time to switch to that API.  Also added
a helper so running a command and capturing output is as simple
as run_command(command).

Fixes the following exception from Popen():
TypeError: __init__() got an unexpected keyword argument 'text'

Signed-off-by: Jeff Mahoney <jeffm@suse.com>