exportpatch: Error out when a commit cannot be located
Error out with an error code when a git commit cannot be found, for
whatever reason, so that scripts using exportpatch can get notified
about the error and can stop processing.

Also, it is better to error out when handling multiple patches than the
error message getting lost in copious output.

Signed-off-by: Borislav Petkov <bp@suse.de>