patchtools: Support commits which come from mainline repo checkout but different remote
Existing code assumes that commits queued in a maintainer's subsystem
will come from a distinct checkout of the subsystem. I.e the case where
1 directory has multiple remotes (mainline and non-mainline) is not
handled correctly. Namely, such commits are erroneosly tagged as if
they are coming from mainline.

Fix this by extending the logic which distinguishes the source of the
commit, not only based on the directory and 'origin' remote but also on
the actual remote where the commit came from. This allows to invoke
exportpatch with a commit id which comes from a maintainer's repo, added
as a remote to the mainline repo i.e no need to do a distinct checkout.

For example:

  $exportpatch 7b804135d4d1 | head -n 5

  From: Kai Huang <kai.huang@intel.com>
  Date: Tue, 15 Aug 2023 23:02:05 +1200
  Subject: x86/virt/tdx: Make TDX_MODULE_CALL handle SEAMCALL #UD and #GP
  Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/tip/tip.git
  Git-commit: 7b804135d4d1f0a2b9dda69c6303d3f2dcbe9d37
  Patch-mainline: Queued in subsystem maintainer repo

Before the output would have been:

  From: Kai Huang <kai.huang@intel.com>
  Date: Tue, 15 Aug 2023 23:02:05 +1200
  Subject: x86/virt/tdx: Make TDX_MODULE_CALL handle SEAMCALL #UD and #GP
  Git-commit: 7b804135d4d1f0a2b9dda69c6303d3f2dcbe9d37
  Patch-mainline: v6.6 or v6.6-rc2 (next release)